Add extra test for do_checkpoint and move DatasetCheckpoint to distilabel.dataset
