+++ tests/test_dataset.py
@@ -16,9 +16,8 @@ import tempfile
 from pathlib import Path
 
 import pytest
-from distilabel.dataset import CustomDataset
+from distilabel.dataset import CustomDataset, DatasetCheckpoint
 from distilabel.tasks import UltraFeedbackTask
-from distilabel.utils.dataset import DatasetCheckpoint
 
 
 @pytest.fixture
@@ -45,13 +44,26 @@ def test_dataset_load_disk(custom_dataset):
         assert isinstance(ds_from_disk.task, UltraFeedbackTask)
 
 
-def test_do_checkpoint():
-    chk = DatasetCheckpoint(save_frequency=2)
-    assert chk.do_checkpoint(0) is False
-    assert chk._total_checks == 0
-    assert chk.do_checkpoint(2) is True
-    assert chk._total_checks == 1
-    assert chk.do_checkpoint(3) is False
-    assert chk._total_checks == 1
-    assert chk.do_checkpoint(4) is True
-    assert chk._total_checks == 2
+@pytest.mark.parametrize(
+    "save_frequency, dataset_len, batch_size, expected",
+    [
+        (1, 10, 1, 10),
+        (3, 10, 1, 3),
+        (8, 32, 8, 4),
+        (8, 64, 16, 4),
+        (20, 100, 7, 5),
+    ],
+)
+def test_do_checkpoint(
+    save_frequency: int, dataset_len: int, batch_size: int, expected: int
+):
+    ds = CustomDataset.from_dict(
+        {"input": ["a"] * dataset_len, "generations": ["a"] * dataset_len}
+    )
+    chk = DatasetCheckpoint(save_frequency=save_frequency)
+    ctr = 0
+    for batch_i, _ in enumerate(ds.iter(batch_size=batch_size), start=1):
+        step = batch_i * batch_size
+        if chk.do_checkpoint(step):
+            ctr += 1
+    assert ctr == expected == chk._total_checks
